+++ src/main/java/nextstep/domain/config/LadderGameConfig.java
@@ -1,4 +1,4 @@
-package nextstep.domain;
+package nextstep.domain.config;
 
 public class LadderGameConfig {
