+++ src/main/java/nextstep/domain/ladder/LadderGameManager.java
@@ -1,9 +1,14 @@
-package nextstep.domain;
+package nextstep.domain.ladder;
+
+import nextstep.domain.track.Tracks;
+import nextstep.domain.line.Line;
+import nextstep.domain.line.Lines;
+import nextstep.domain.person.Participants;
 
 import java.util.Collections;
 import java.util.List;
 
-import static nextstep.domain.LadderGameConfig.TRACK_STEP_SIZE;
+import static nextstep.domain.config.LadderGameConfig.TRACK_STEP_SIZE;
 
 public class LadderGameManager {
