+++ src/main/java/nextstep/domain/ladder/LadderResultFactory.java
@@ -1,4 +1,4 @@
-package nextstep.domain;
+package nextstep.domain.ladder;
 
 import java.util.Arrays;
 import java.util.List;
