+++ src/main/java/nextstep/domain/line/strategy/NotAllDrawTransverseBarStrategy.java
@@ -1,17 +1,17 @@
-package nextstep.domain;
+package nextstep.domain.line.strategy;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class AllDrawTransverseBarStrategy implements TransverseBarStrategy {
+public class NotAllDrawTransverseBarStrategy implements TransverseBarStrategy {
 
-    private static final boolean ALL_DRAW = true;
+    private static final boolean NOT_ALL_DRAW = false;
 
     @Override
     public List<Boolean> draw(int countOfPerson) {
         return IntStream.range(0, countOfPerson - 1)
-                .mapToObj(count -> ALL_DRAW)
+                .mapToObj(count -> NOT_ALL_DRAW)
                 .collect(Collectors.toList());
     }
