+++ src/main/java/nextstep/domain/person/Person.java
@@ -1,4 +1,4 @@
-package nextstep.domain;
+package nextstep.domain.person;
 
 import java.util.Objects;
