+++ src/test/java/nextstep/domain/LadderResultTest.java
@@ -1,5 +1,6 @@
 package nextstep.domain;
 
+import nextstep.domain.ladder.LadderResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
