+++ src/test/java/nextstep/domain/ParticipantsTest.java
@@ -1,11 +1,10 @@
 package nextstep.domain;
 
-import nextstep.view.InputView;
+import nextstep.domain.person.Participants;
+import nextstep.domain.person.PersonFactory;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
