+++ src/test/java/nextstep/domain/PersonTest.java
@@ -1,5 +1,6 @@
 package nextstep.domain;
 
+import nextstep.domain.person.Person;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
